Second attempt to support bedrock skins
This time the bedrock skin support should be fully OK. Thanks to UUID stored by mcMMO we can easily know if a player is a bedrock player or not. If it is, then we use the TydiumAPI with his UUID to display his skin. Otherwise we still use player name to show his skin.